feature(ci): Add clang-format check and unit test support
This commit adds two github action for CI. The first is to use `cmake`
to build the source code and build the document and tests the unit test.

Also this commit makes sure that we use `clang-format` to ensure the
consistent of the code style. So uses `jidicula/clang-format-action@v4.6.2`
to check the code style.